backend/pyproject.toml
[tool.poetry]
name = "src"
version = "1.0.0"
description = "API for school app"
authors = ["Kennedy Uche <chukwukakennedyuche@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
passlib = "^1.7.4"
bcrypt = "^4.1.2"
cryptography = "^40.0.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.7"
starlette = "^0.26.1"
databases = "0.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"